Rework file upload to have Dropzone and from url
I created two different forms for the Dropzone and the url
as it was easier to handle that way. That meant that I
had to add the AuthenticityToken in to both forms, which
is a little ugly.

I reckon this could & should be reworked to use a framework
like vue-js at somepoint. It would make reasoning about the
code a lot easier.

The reset-css also makes styling the entire thing a little
bit more difficult.