Fix bug with xfigure insert on upload.
If the xfigure is inserted on upload then the callback does not work as
it's targeting the dialog environment, when actually its in a popup.

This "fixes" it (hack) so that we just go back to the browse screen and
can choose from there. It should be fixed properly at some point, but,
I'm not giving more time to it now.