Basics of word-style dropdown added.
Adds the Dropdown menu in under the wordstyle button (for the moment,
this is really a bit naff)

I've also altered the properties so that we just use a single js include
for all the options. We don't want the various shared options anyway so
this makes sense. I think really this dialog could be reworked into Vue
or _something_ with a bit more validation around it.

That said, it's working.. so.. :)