Merging the justify options.
I would _love_ to be able to say that this was my cleverness, but, it
totally wasn't.

Liberally stolen solution from StackOverflow:

https://stackoverflow.com/questions/44283793/ckeditor-4-7-justify-group-menu-button

Only issue is that the menu icon doesn't change to show the style of the
menu which is currently in use. Which may be a blocker I'm not sure.