Adds the dropzone (list and table) elements
One thing which I did change was to remove out of the eastableproperties
the _each method which was (I think) supposed to be saving off the
properties correctly.

It didn't seem like that was actually working properly, especially for
booleans, so, I removed it.

It all seems to continue to work correctly.. so... yeah, not sure.

Maybe I have broken the advanced properties though, I'll check with Zack