Add dropzone properties to the table
The table needs to to be able to set whether it is for dropzones, and if
so, what the column and position values are?!

We need two keys added for tables:  "dropnum" and "dropcols".  I'll explain what they do later.  For now, you can add them to the advanced (right-click) dialog.  Also, some variant should be added to the new table properties tab that you made.

Both of the new keys take non-negative integer values.  "Auto" gets passed to TeX as the number 0.  Otherwise, the range for drop positions is 1-20 and drop columns 1-10.

Also add two keyval properties to the drag-items list advanced.